improve error handling in future builders
still need to find a way to show a toast; if I try to call it on
snap.hasError, I get:
setState() or markNeedsBuild() called during build. This Overlay widget cannot be marked as needing to build because the framework is already in the process of building widgets. A widget can be marked as needing to be built during the build phase only if one of its ancestors is currently building