Fix default relevance picker not being set.
The default relevance should automatically be set to the
user's group or the global filter settings. We probably
broke it when we started saving the filter to Firebase.

I fixed it by fetching the filter in Firebase by default
(even for non-global filters). It's still not perfect,
sometimes in the add event page nothing shows in the
relevance picker and I'm not sure why, but if you press
on Custom, it's automatically set to the current filter
options.