Started the implementation for default semester
The default semester shown during editing or during the creation of a new event is initialized with the one closest to today, I will rewrite it soon using a more readable method, like "contains".

Then we check if the user clicked an existing event or a empty space on the calendar. If it already exists, we can extract the semester from the details. If the user wants to create a new one, we iterate through the calendars, then through their semesters and check whether the event is before or during that certain semester.