Displaying the correct semester when editing / creating an event (#135)
* The default semester shown during editing or during the creation of a new event is initialized with the one closest to today, I will rewrite it soon using a more readable method, like "contains".

Then we check if the user clicked an existing event or a empty space on the calendar. If it already exists, we can extract the semester from the details. If the user wants to create a new one, we iterate through the calendars, then through their semesters and check whether the event is before or during that certain semester.

* Fixed displayed semester for new events

possible todo - default value for semester right at the beginning of the state initialization method

* made a non-reassigned value final, removed unnecesary comments

* decided to use isDuring instead of isBeforeOrDuring

that means that, when choosing what semester to display, I take into account whether that event is part of that semester

* changed back to isBeforeOrDuring, with reverse iterating

iterating from the latest to the earliest semester ensures that we choose the semester which is closest to the event (and even potentially after it)

* fixed semester iteration, added edge cases (holidays)

* fixed semester iteration, added edge cases (holidays)

* added a comment to explain the parsing of the semester id, simplified a return value

* removed unnecessary new line

* Apply suggestions from code review

Co-authored-by: Ioana Alexandru <ioanaa.alexandru98@gmail.com>

* clarified semester comment

* Fix bugs when adding new info to a class. (#131)

* Fix bugs when adding new info to a class.

Before, it would break if trying to add info to a
class that didn't yet have a document associated,
or if the grading field was missing.

* Bump version.

* Fix default relevance picker not being set. (#130)

* Fix default relevance picker not being set.

The default relevance should automatically be set to the
user's group or the global filter settings. We probably
broke it when we started saving the filter to Firebase.

I fixed it by fetching the filter in Firebase by default
(even for non-global filters). It's still not perfect,
sometimes in the add event page nothing shows in the
relevance picker and I'm not sure why, but if you press
on Custom, it's automatically set to the current filter
options.

* Bump version.

* fixed comment about semester.id indexing

* Apply suggestions from code review

Co-authored-by: Ioana Alexandru <ioanaa.alexandru98@gmail.com>

* increased build number

Co-authored-by: Ioana Alexandru <ioanaa.alexandru98@gmail.com>