Fixed bug which prevented GCal imported events from being edited. Organized code
My guess is that the aforementioned bug was caused by RRule inconsistencies. According to RFC5545 (https://tools.ietf.org/html/rfc5545) [BYWEEKNO] MUST NOT be used when the FREQ rule must not be used when the FREQ part is set to anything other than YEARLY.', so I replaed 'Daily' with 'Yearly'. On my test events, things seem to work well, but I will further investigate this.